+++ src/I18nContext.js
@@ -1,6 +1,6 @@
 // @flow
 import * as React from 'react';
-import { type I18nType } from '.';
+import { type I18nType } from './types';
 import mockI18n from './mockI18n';
 
 const I18nContext: React.Context<I18nType> = React.createContext(mockI18n);
